Import files provided by Microchip FPGA support
This is the project provided to me by Microchip FPGA support. It can be used to demonstrate how to reconstitute a Libero project that uses a PolarFire FPGA. All comments and readme files in this are as provided by Microchip. I do not offer any support for this whatsoever. I just provide it to help others who are looking for a starting point for how to use tcl with Libero.